# Appraisal-Teaching-Processes/dict/models.ts
/* eslint-disable */
// tslint:disable
/**
 * Appraisal Service
 * Appraisal Service
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

export namespace Api {
	export interface MessageData {
		message: string;
	}

	export interface TeachingData {
		siNo?: string;
		course?: string;
		level?: string;
		modeOfTeaching?: string;
		/**
		 * @type {number}
		 * @memberof TeachingData
		 */
		classesAllotted?: number;
		/**
		 * @type {number}
		 * @memberof TeachingData
		 */
		classesConducted?: number;
		facultyName?: string;
	}

	export interface TeachingPagedResultData {
		/**
		 * @type {number}
		 * @memberof TeachingPagedResultData
		 */
		totalCount?: number;
		items?: Api.TeachingData[];
	}

}